requirements.txt
setuptools~=65.5.0
pip~=23.1.2
pytz~=2023.3
Django~=4.2.2
asgiref~=3.7.2
sqlparse~=0.4.4
Markdown~=3.4.3
django-filter~=23.2
djangorestframework~=3.14.0
Jinja2~=3.1.2
dill~=0.3.6
isort~=5.12.0
colorama~=0.4.6
requests~=2.31.0
flake8~=6.0.0
pyflakes~=3.0.1
pycodestyle~=2.10.0
MarkupSafe~=2.1.3
pylint~=2.17.4
astroid~=2.15.6
tomlkit~=0.11.8
mccabe~=0.7.0
platformdirs~=3.8.1
wrapt~=1.15.0
certifi~=2023.5.7
itypes~=1.2.0
coreapi~=2.3.3
coreschema~=0.0.4
uritemplate~=4.1.1
idna~=3.4
urllib3~=2.0.3